+++ torch/fx/experimental/validator.py
@@ -5,7 +5,7 @@ import operator
 import sympy
 
 from dataclasses import dataclass
-from typing import Any, Callable, Dict, Iterable, Set, Tuple, Type, Union
+from typing import Any, Callable, Dict, Iterable, List, Set, Tuple, Type, Union
 
 import torch
 from torch._dynamo.exc import TorchDynamoException
@@ -58,6 +58,9 @@ try:
     def z3str(e: z3.ExprRef) -> str:
         assert z3.is_expr(e), f"unsupported expression type: {e}"
 
+        def get_args_str(e: z3.ExprRef) -> List[str]:
+            return [z3str(e.arg(i)) for i in range(e.num_args())]
+
         # First, we simplify the given expression.
         # This is done using rewriting rules, so shouldn't take long.
         e = z3.simplify(e)
@@ -74,17 +77,14 @@ try:
         decl = e.decl()
         kind = decl.kind()
         op = str(decl)
-        args = [z3str(e.arg(i)) for i in range(e.num_args())]
+        args = get_args_str(e)
 
-        if kind == z3.Z3_OP_DISTINCT:
-            op = "!="
-
-        elif kind == z3.Z3_OP_POWER:
+        if kind == z3.Z3_OP_POWER:
             op = "pow"
 
-        elif kind in (z3.Z3_OP_ADD, z3.Z3_OP_SUB, z3.Z3_OP_MUL):
-            # Collect the arguments of chains of ADD, SUB, and MUL.
-            # This is safe, since they are commutative.
+        elif kind in (z3.Z3_OP_ADD, z3.Z3_OP_MUL):
+            # Collect the arguments of chains of ADD and MUL.
+            # This is safe, since they are associative.
 
             def collect_str_args(e):
                 if not (z3.is_app(e) and e.decl().kind() == kind):
@@ -97,6 +97,28 @@ try:
                     ]
 
             args = collect_str_args(e)
+
+        elif kind == z3.Z3_OP_NOT:
+            # Revert some conversions that z3.simplify applies:
+            #   - a != b ==> (Not (== a b)) ==> (!= a b)
+            #   - a < b ==> (Not (<= b a)) ==> (> b a)
+            #   - a > b ==> (Not (<= a b)) ==> (> a b)
+
+            assert e.num_args() == 1
+            arg = e.arg(0)
+
+            assert z3.is_app(arg)
+            argkind = arg.decl().kind()
+
+            logic_inverse = {
+                z3.Z3_OP_EQ: "!=",
+                z3.Z3_OP_LE: ">",
+                z3.Z3_OP_GE: "<",
+            }
+
+            if argkind in logic_inverse:
+                op = logic_inverse[argkind]
+                args = get_args_str(arg)
 
         elif kind in (z3.Z3_OP_TO_INT, z3.Z3_OP_TO_REAL):
             assert e.num_args() == 1
